requirements.txt
laktory
openai
pillow
pytesseract
